started on the routing and setting up other pages
